refactor: add extension function and format values

+++ src/main/kotlin/org/example/gamebusters/model/Plan.kt
@@ -1,9 +1,11 @@
 package org.example.gamebusters.model
 
+import java.math.BigDecimal
+
 sealed class Plan(val type: String) {
 
-    open fun getValue(rent: Rent): Double {
-        return rent.game.price * rent.rentalPeriod.inDays
+    open fun getValue(rent: Rent): BigDecimal {
+        return rent.game.price * rent.rentalPeriod.inDays.toBigDecimal()
     }
 
 }
